Initial layout: tri-layer with OLED shenanigans
I've now made enough progress on this layout that I feel that it's worth
committing. So far, I have a tri-layer layout heavily optimised for
programming in C, C++, and Rust, with most languages being generally
well supported. In addition to the standard layout, I also have a
navigation layer that is definitely a work in progress, and a
hexadecimal numpad to stave off the urge to make any further hardware
purchases.

On the aesthetic front, I've gutted the default keyboard definition
since it pollutes IMEM with logos and other nonsense for which I have no
use. In its stead, I've added a counter on the left side (more to come
in future commits, undoubtedly); on the right are a keystroke counter
and a dynamic Bongo Cat animation, which is the single largest use of
IMEM in this layout. Both numeric values are rendered as Cistercian
numerals, both because they are easier to render than decimal digits and
because they look amazing.

This commit message was typed on this keyboard, using the current
version of this layout. :)